% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2granges.R
\name{bed2granges}
\alias{bed2granges}
\title{bed to GRanges}
\usage{
bed2granges(df)
}
\arguments{
\item{df}{bed3 like dataframe}
}
\value{
a GRanges object
}
\description{
This function converts bed3-like format object to a GRanges object
}
\keyword{DMR}
\keyword{GRanges}
\keyword{Promoter}
