% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMREnhancer.R
\name{DMREnhancer}
\alias{DMREnhancer}
\title{Extraction of DMR associated enhancer}
\usage{
DMREnhancer(sig_targets_gr)
}
\arguments{
\item{sig_targets_gr}{GRanges object}
}
\value{
a GRanges object of DMR horboring enhancers
}
\description{
This function extract a Enhancer horboring DMR as a GRanges object
}
\keyword{DMR}
\keyword{Enhancer}
\keyword{GRanges}
