% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_gr.R
\docType{data}
\name{gencode_gr}
\alias{gencode_gr}
\title{Genecode GRanges object}
\format{A GRanges object}
\usage{
gencode_gr
}
\description{
This data comes from Genecode database (comprehensive gene annotation hg19),
\url{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_19/gencode.v19.annotation.gff3.gz}. 
Leaving "gene", other rows were removed.
the gff3 format file was loaded by R as gencode_gr by gff32granges.
}
\examples{
gencode_gr
}
