% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff32granges.R
\name{gff32granges}
\alias{gff32granges}
\title{gff3 to GRanges}
\usage{
gff32granges(gff3file)
}
\arguments{
\item{gff3file}{location of a gff3 file}
}
\value{
a GRange object
}
\description{
This function loads gff3 format file and store it as a GRanges object
}
\keyword{GRanges}
\keyword{gff3}
