% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_bed.R
\docType{data}
\name{gh_bed}
\alias{gh_bed}
\title{GeneHancer bed-like format object}
\format{A bed-like object}
\usage{
gh_bed
}
\description{
This data comes from GeneHancer database,
\url{https://academic.oup.com/database/article/doi/10.1093/database/bax028/3737828}. 
The data are integrated human enhancer data set and gene connection of each enhancer.
The original data is downloaded from
\url{https://genecardsdata.blob.core.windows.net/public/GeneHancer_version_4_4.xlsx}
The header was removed and stored as tab-deleted gff3-like format file.
The gff3-like format file was forther converted to a  bed format file, then liftover from hg38 to hg19.
the hg19 bed format file was loaded by R as gh_bed.
}
\examples{
gh_bed
}
