% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDissociate.R
\name{metaDissociate}
\alias{metaDissociate}
\title{extraction of metadata infomation}
\usage{
metaDissociate(metadata)
}
\arguments{
\item{metadata}{semi-colon deleted metadata}
}
\value{
a data frame of links between enhancers and connected genes
}
\description{
This function dissociate metadata of GRanges class of gff3 files
}
\keyword{gff3}
\keyword{metadata}
