% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metTFNetmulti.R
\name{metTFNet.multi}
\alias{metTFNet.multi}
\title{Target network of multiple methyl-regulationg TF}
\usage{
metTFNet.multi(MethylDemethyl = "Demethyl",
  eg_connection_list = eg_connection_list,
  sig_genes_list = sig_genes_list, outname = "")
}
\arguments{
\item{MethylDemethyl}{a vector. methylation or demethylation mediated by the methyl-regulating TF. "Methyl" or "Demethyl"}

\item{eg_connection_list}{a list of connection data between enhancers of the methyl-regulating TF targets and thir downstream gene. the data have to have score at 3rd column.}

\item{sig_genes_list}{a list of vectors of the methyl-regulating TF target promoters}

\item{outname}{name for output file}
}
\value{
a dataframe of connections (and export a network plot as a pdf)
}
\description{
This function draws a DNA methylation regulatory network conposed of methyl-regulating TF, target promoter, target enhancer, and enhancer connected genes
This fanction is designed for multiple methyl-regulating TF input.
output "TF_methyl_network.pdf" and connection dataframe
}
\keyword{enhancer}
\keyword{methylation}
\keyword{network}
\keyword{promoter}
