% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metTFNet.R
\name{metTFNet}
\alias{metTFNet}
\title{Target network of methyl-regulationg TF}
\usage{
metTFNet(TF = "TF", MethylDemethyl = "Demethyl",
  enhancer_gene_connection = enhancer_gene_connection,
  sig_genes = sig_genes, outname = "")
}
\arguments{
\item{TF}{a vector. name of a methy-regulating TF}

\item{MethylDemethyl}{a vector. methylation or demethylation mediated by the methyl-regulating TF. "Methyl" or "Demethyl"}

\item{enhancer_gene_connection}{connection data between enhancers of the methyl-regulating TF targets and thir downstream gene. the data have to have score at 3rd column.}

\item{sig_genes}{a vector of the methyl-regulating TF target promoters}

\item{outname}{name for output file}
}
\value{
a dataframe of connections (and export a network plot as a pdf)
}
\description{
This function draws a DNA methylation regulatory network conposed of methyl-regulating TF, target promoter, target enhancer, and enhancer connected genes
 output "TF_methyl_network.pdf" and connection dataframe
}
\keyword{enhancer}
\keyword{methylation}
\keyword{network}
\keyword{promoter}
