% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotPosiList.R
\name{MotPosiList}
\alias{MotPosiList}
\title{Motif position list at DMRs}
\usage{
MotPosiList(infile = "sel_processed_Mval.txt", motifDBList,
  ControlColnum, TreatmentColnum, p.cutoff = 0.001, cutoff = 2,
  MethylDemethyl = "Demethyl", version = "850", sampling = FALSE,
  seq_range = c(-200, 200), outname = "motif_num_dist",
  nbiom_cutoff = 0.05)
}
\arguments{
\item{infile}{input m-value file}

\item{motifDBList}{list of motif PWM}

\item{ControlColnum}{control column index(s)}

\item{TreatmentColnum}{treatment column index(s)}

\item{p.cutoff}{cutoff for multi-sample comarison (p-value)}

\item{cutoff}{cutoff for single sample comparison (delta M)}

\item{MethylDemethyl}{Metylation or Demethylation analysis}

\item{version}{version of methylation array 450 or EPIC(850)}

\item{sampling}{number of sampling. if FALSE, use all data}

\item{seq_range}{range from CpG to be analyzed}

\item{outname}{output file name}

\item{nbiom_cutoff}{cutoff of negative binomial model test}
}
\value{
a List of lists of motif positions in DMR from CpG sites and of that in random regions
}
\description{
This function performes differentially methylation analysis and motif enrichment analysis at the DMR
}
\keyword{DMR}
\keyword{Infinium}
\keyword{TF}
\keyword{motif}
