% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsplotr.R
\name{ngsplotr}
\alias{ngsplotr}
\title{reads enrichment plots}
\usage{
ngsplotr(bam.files, peaks, bin.num = 100, sample_names,
  range = c(-5000, 5000), outname = "my_enrichment")
}
\arguments{
\item{bam.files}{a vector of bam.file locations}

\item{peaks}{a GRanges object for regions to be analyzed.}

\item{bin.num}{number of bin for score matrix}

\item{sample_names}{sample names of bam files}

\item{range}{a raneg to be analyzed}

\item{outname}{name for output file}
}
\value{
a scorMatrix object and export a line plot, heatmaps(k-mean and hierarchical clusterings) as a pdf)
}
\description{
This function draws a line plot foraverage enrichment of  ngs reads and 
heatmaps for ngs reads enrichment (k-means and hierarchical clusterings).
}
\keyword{ChIP-seq}
\keyword{bam}
\keyword{enrichment}
\keyword{heatmap}
\keyword{ngs}
\keyword{ngsplot}
