% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptRange.R
\name{OptRange}
\alias{OptRange}
\title{Number of significant targets}
\usage{
OptRange(infile = infile, motifDBList = motifDBList,
  ControlColnum = ControlColnum, TreatmentColnum = TreatmentColnum,
  MethylDemethyl = MethylDemethyl, version = version,
  outname = outname, nbiom_cutoff = nbiom_cutoff, ranges)
}
\arguments{
\item{infile}{input m-value file}

\item{motifDBList}{list of motif PWM}

\item{ControlColnum}{control column index(s)}

\item{TreatmentColnum}{treatment column index(s)}

\item{MethylDemethyl}{Metylation or Demethylation analysis}

\item{version}{version of methylation array 450 or EPIC(850)}

\item{outname}{output file name}

\item{nbiom_cutoff}{cutoff p-value for negative binomial test}

\item{ranges}{a vector. ranegs to be tested.}
}
\value{
a matrix of results of fitting and negative binominal test
}
\description{
This function compute the number of significant targets of methyl-regulating TFs in differentranges.
Number of motifs in a region is dependent of the rength of the region. 
To potimize the range of the region, this function return the number of significatn targets changing the range.
Becaue fitting function is not robust to outliers, if a fitting return the error, repret the fitting by 100 times.
output figures consist of histgram of motif number at DMR and at random regions, fitting model line plot for each results.
and return tabel of the result
}
\keyword{binominal}
\keyword{fitting}
\keyword{optimization}
